+++ src/Cassandra/Requests/OptionsRequest.cs
@@ -25,7 +25,7 @@ namespace Cassandra
 
         }
 
-        public RequestFrame GetFrame(byte streamId, byte protocolVersionByte)
+        public RequestFrame GetFrame(short streamId, byte protocolVersionByte)
         {
             var wb = new BEBinaryWriter();
             wb.WriteFrameHeader(protocolVersionByte, 0x00, streamId, OpCode);
