+++ src/Cassandra/Requests/RegisterForEventRequest.cs
@@ -35,7 +35,7 @@ namespace Cassandra
                 _eventTypes.Add("SCHEMA_CHANGE");
         }
 
-        public RequestFrame GetFrame(byte streamId, byte protocolVersionByte)
+        public RequestFrame GetFrame(short streamId, byte protocolVersionByte)
         {
             var wb = new BEBinaryWriter();
             wb.WriteFrameHeader(protocolVersionByte, 0x00, streamId, OpCode);
